* Fix compile error for #define UNAME_HACK * Fix crash when you do $open(/file/doesnt/exist R) * Add an /on 338 default handler. * Fix integer underflow with log file refnums * Rename STATUS_USER0 to STATUS_USER for backwards compat. oops! * Whack BRACE_LOAD_HACK, it's been unsupported for a while.  PF loader rules! * Don't allow /botmode if #define NO_BOTS * Fix DCC connections instantly "timing out"? (Check on this) * Whack FLOATING_POINT_SUPPORT and depend only on /set floating_point_math

+++ source/logfiles.c
@@ -1,4 +1,4 @@
-/* $EPIC: logfiles.c,v 1.28 2004/11/02 23:17:34 jnelson Exp $ */
+/* $EPIC: logfiles.c,v 1.29 2005/01/13 16:06:06 jnelson Exp $ */
 /*
  * logfiles.c - General purpose log files
  *
@@ -683,7 +683,7 @@ BUILT_IN_COMMAND(logcmd)
 }
 
 /****************************************************************************/
-void	add_to_logs (int winref, int servref, const char *target, int level, const char *orig_str)
+void	add_to_logs (long winref, int servref, const char *target, int level, const char *orig_str)
 {
 	Logfile *log;
 	int	i;
